Manual

EquiBaryInterp.BaryPoly — Type

BaryPoly(x, y)
BaryPoly(x, y, w)

Constructs a barycentric Lagrange polynomial from the data y sampled on x. (See barycentric_weights.)

EquiBaryInterp.LocalEquiBaryInterp — Type

LocalEquiBaryInterp(x::AbstractVector, y::AbstractVector, [degree=8])
LocalEquiBaryInterp(x::Vector, y::Vector, w::Vector, h⁻¹)

Construct a local barycentric Lagrange interpolant that forms a degree degree local polynomial approximation of the data y on the equispace grid x, which must be identical to a sorted range with step size h. w are the equispace interpolation weights (see equi_bary_weights). The restrictions on x are that it be a real vector with entries identical to a range and that it be sorted. y can contain any type supporting addition and scalar multiplication.

Returns an interpolant f that can be evaluated at points within extrema(x) using a function syntax f(x).

EquiBaryInterp.bary_kernel — Method

bary_kernel(x, xs, ys, ws)

Computes the value of the barycentric Lagrange polynomial interpolant with nodes xs, values ys, and weights ws at point x.

EquiBaryInterp.barycentric_weights — Method

barycentric_weights(x::AbstractVector{T}) where {T<:Real}

Computes barycentric weights for the nodes x.

EquiBaryInterp.equi_bary_weights — Method

equi_bary_weights(n::Integer)

Computes barycentric weights for n+1 equispace nodes.

EquiBaryInterp.to_sorted_range — Method

to_sorted_range(x::AbstractVector)

Assert that x is numerically identical to an equispace range, and return an equivalent range object.